move slashcommand listener from index.js to events folder
